[MDC-1010] simet-api: add report queue subsystem
Add a reports queue processing system.

simet_lmap-send-report.sh (the "report sink" as far as the lmapd
schedule is concerned) will compose/render the report into the reports
queue, and try to send it immediately.  Should that fail, the report
will be compressed and left in the queue for later sending.

simet_report-queue-flush.sh locates unsent reports in the queue, and
tries to send them all.  It should be run periodically.  The queue can
have subdirectories, if needed (this will be used later).

It also cleans up stale data (more than 24h old) in the queue that
doesn't match a compressed report filename.

simet_lmap-send-report.sh changed the behavior re. errors returned by
"lmapctl report": since version 0.15.0 of simet-lmapd, it will return a
non-zero status only if it cannot render *any* report at all, either
because *all* the input data was invalid, or the lmap subsystem
configuration is broken.  So, we now always return status zero if
we got that far, to erase likely invalid input data.