[MDC-1010] simet-api: add a spool directory for unsent reports
Switch the scripts and config to work with the base lmap spool dir,
and define LMAP_SPOOL_DIR in simet-ma.conf.  We keep LMAP_QUEUE_DIR
in there, though.

Create a new spool sub-directory for unsent reports.

Note that we do not substitute @LMAPQUEUEDIR@ in .in files anymore,
hardcoding should be on @LMAPSPOOLDIR@/queue (or /reports), or use
the simet-ma.conf defines.