[MDC-1010] simet-api: more control for report submission queues
Add CLI options to simet_lmap-send-report to bypass queueing or sending
of a report, and to give up faster when the LMAP report collector is
slow to respond.

Add a CLI option to give up faster on the LMAP report collector to
simet_report-queue-flush.sh, for parity with simet_lmap-send-report.sh.

Change the LMAP task configuration for "simet.nic.br_report" to not wait
too much for the LMAP report collector (and queue the report for later
transmission, instead).