[MDC-1010] simet-api: sort report queue by time
When generating the report, add the current time to its filename
and use that to try to send them in the correct order later.

The expected situation where the system time changes is of little
consequence: it usually will go forward (which preserves the desired
ordering), and we are not really supposed to care much about the
ordering in the server-side pipeline and data stores anyway.