[MDC-1010] simet-api: remove unsent reports when low on resources
If we're low on resources (defined as free filesystem space in the
filesytem hosting the report queue directory for the moment, and
"MemFree" RAM if a tmpfs is detected in Linux), remove unsent reports
until we have freed enough resources.

The queue is sorted by filename in ascending order, and files are
deleted in that order until enough resources have been freed.

For now, consider we shouldn't bother removing old reports when we have
more than 100MiB free space in that filesystem, or if we're using less
than 100KiB in the report queue, unless we're quite short on MemFree.

Rework the command line processing to accept options, and add two
options: --max-age (to remove unsent reports by age, in days) and
--max-size, to remove unsent reports when the whole unsent queue is
above a size threshold, in kibibytes.

Note that --max-age *can* and *will* purge recent reports that we'd
prefer to retain due to lack-of-RTC mishaps on embedded, so it is not
applied when there are enough free resources.

--max-size is implemented as a "soft" limit right now: if it'd increase
available filesystem space by less than 10%, it is not applied (but the
other safety thresholds described above can still cause unsent data to
be removed).