[MDC-1010] simet-api: try to send reports more often (SIMETBox)
On SIMETBox, try to render pending reports and send them every five
minutes.  Unsent reports are compressed and will wait for a queue flush
(every 15 minutes).

On non-SIMETBox, we already try to render pending reports every minute,
but that might disturb a low-power single-core CPU too much.

This is a stopgap solution: we want to render and either send or
compress these reports as soon as possible, and that calls for a
different solution.