[MDC-1026] simet-api,deb: simet_lib*: defend against config syntax errors
Unless we go to extreme lengths, it is quite hard to ensure the shell
won't abort on us when sourcing the configuration files.  We can try,
though...

This change requires some updates to the packaging scripts that call
simet_lib-config.sh, otherwise they'd not benefit any from the change
(but it wouldn't break them, either).