deb: postinst,postrm: cosmetic fixes
Add a comment to the /etc/cron.d/simet-ma autogenerated file to that
effect.  Note that we don't overwrite that file right now if it already
exists, so this change is only for new installs right now.

Don't try to remove a conffile dpkg will already remove for us in
postrm (there are other conffiles we are not removing like that, being
consistent is important), and add a comment about config files like
/etc/cron.d/simet-ma needing the explicit rm.

Cleanup whitespace in postrm.