jsonpath/libubox: update embedded copy to latest git
simet-ma 'vendors' a copy of jsonpath and libubox to be used when the
local system lacks jsonpath and/or libubox (which is almost always the
case outside of openwrt).

Update the libubox component to the latest git HEAD:
libubox commit 551d75b5662cccd0466b990d58136bdf799a804d

Note: this patch removes any local changes.  We will apply our local
patches to libubox in the next commit in this series.