jsonpath/libubox: refresh and apply local changes
Apply local changes to the libubox update, and refresh the patch to
remove fuzz.

We need to also rename libubox's "assert.h" out of the way, because it
ends up in the include path of jsonpath, and breaks its build.

Note: switch from "diff -ru" to "git diff --no-renames" to generate the
patch file.