[MDC-1080] tcpbw: try to request very large socket buffers
Ask for 32MiB worth of send and receive socket buffers before connect,
so that TCP window scale can handle LFNs (think 100Mbps @ 500ms RTT).

Otherwise, the number of concurrent connections required to saturate
a LFN (high-bandwidth, high-delay network) is too high.

The way we do it is not pretty, since we directly modify the forked
libubox.  This actually asks for some refactoring to get rid of the
libubox layer...

OTOH, this means that right now we won't change anything in SIMETBox
builds, where the openwrt native libubox is used.