[MDC-1080] tcpbw: use correct ssize_t type for recv() result
Fix (or defang?) an issue in receiveDownloadPackets(), by using the
correct signed type (ssize_t) to catch the result of recv() calls.

Without this fix, should select() ever let the recv() codepath run in a
condition where it would error out, it would not detect the error due to
size_t(-1) > 0.

Should this bug trigger, it would cause size_t(-1) to be added to the
total bytes downloaded for that stream, with a net effect of decreasing
it by one in most platforms, where sizeof size_t == sizeof uint64_t, or
causing worse measurement errors in in platforms where these sizes
differ.

This issue has been observed in the field, and results in the data
pipeline rejecting the measurement.