[MDC-1080] tcpbw: during download measurement, drain stream before next
We would round-robin doing recv() among the read streams based on
select() triggering for reads.

It was done that way to promote a bit more fairness between streams, but
field testing shows that it can actually worsen aggregate peer TCP
output when the streams are already highly unbalanced and there is
enough throughput for some of them to actually fill (or come close to
filling) their receive buffers.

Change the code to drain the buffer of each stream by doing recv() in a
loop before switching to the next ready stream, instead.  It seems to
help.

We don't drain to EAGAIN/EWOULDBLOCK: we assume this would happen in the
next recv() as soon as it returns a short read, and switch to the next
stream.  This saves us a recv() syscall per loop in the hot path, and
should not cause any trouble in the (in this code, supposedly) rare case
of shorter TCP recv() reads due to signals or URG data.