inetup: use ssize_t to handle recv() result
Use ssize_t to handle the return of recv().  The old code was safe, as
recv() semanthics ensure the results were bound to the int range, but
using the correct types result in code that is just as clean, much more
explicit, and squashes a few warnings.

While at it add a comment documenting one related warning to be a
false-positive.