[MDC-1090] inetup: allow more than one server name
Change the command line server name specification from <host> [<port>]
to <host>[:<port>], and allow more than one server name.

For now, we try to connect to one IPv4 and one IPv6 address of each
specified server name.

This is a backwards-incompatible change to the command line syntax, but
we don't use the optional port anywhere in other glue code.