[MDC-1090] inetup: accept new parameters for uptime group
Accept three new server parameters, and display them in the protocol
logging.  Currently they don't have further use, but we will need them
later.

"uptime-group" is going to be used when measuring uptime to several
different locations (such as IXPs).

"server-description" will aid debugging and problem tracking, and
uniquely identifies a server to the server-team.

"server-hostname" is required due to some central pipeline oddities in
the way it detects MP location.  It will be required if the client
starts generating the LMAP reports itself, and it can be useful for UI
and logging/troubleshooting purposes as well.

These parameters are local to a connection, and can be reset/changed
during a reconnection.

Note: once set, a server should not change its uptime group *except*
through a reconnection.  The client will ignore any attempts to change
it once it was set for a connection.