[MDC-1090] inetup: use nonblocking connect()
Refactor the code and switch to non-blocking connect(), which is much
harder than it should be, due to some quirks of the POSIX/SuSv4 socket
API.

Basically, it is insanely broken unless you are doing things only for
(non-ancient) Linux kernels.  Try to already write the code to be easier
to port for other kernels, and attempt to detect corner cases in
multiple ways.

Without this fix, we would starve the other protocol connections while
one was stuck in blocking connect(), causing inetup protocol timeouts
most likely.

Notes: we depend on setsockopt(SOL_SOCKET, SO_SNDTIMEO/SO_RCVTIMEO) to
adjust the TCP connect timeout.  This could be changed to a manually
controlled connect() timeout, by track the time we're waiting for
connect() to complete and doing a close() if it takes too long.