[MDC-1090] inetup: plug memory leak when reconnecting
When we reconnect to a measurement peer, we'd leak the past contents of
the peer_name/port and local_name/port through xx_nameinfo().

Change xx_nameinfo() to dealocate previous uses of such fields.  While
at it, detect when the field contents did not change, and skip the
free()/strdup() in that case, thus reducing malloc() churn.

Also, change AF_UNSPEC port (actually, service) to "unknown" instead of
"error", to match AF_UNSPEC hostname.

Thanks to valgrind for spotting the leak.