[MDC-882] inetup: measurements: correct timestamp field names
Use "microseconds" instead of "microsseconds" as the field names in
timestamps for measurements in MSG_MEASUREMENT.

This is an incompatible protocol change -- although easily contained and
worked around server-side by accepting both spellings in the inetup
server, and converting it to whatever the pipeline wants.

Currently, the server does not forward measurement messages from the
client anywhere: they're accepted, accounted for, and dropped.  There
isn't anything to change server-side right now.

An appropriate comment was added to the server source code about the
need to do the "microsseconds" -> "microseconds" spelling conversion
should supporting outdated clients still be relevant when the server
side starts being developed.