[MDC-1090] inetup: add a global sequence number to some messages
Add a global sequence number, which is strictly monotonic and shared
among all connections of the client.  Send it on CONNECT (important!)
messages, as well as MEASURMENT and EVENTS (just in case they're later
needed).

The client-provided sequence numbering in the CONNECT message helps the
server to uniquely identify to the pipeline each client connection, even
when they happen on the same instant to different servers.  It also
helps the pipeline detect (large) server clock skew.

It is important that the same sequence number pool (which must be
strictly monotonic and increasing) be used for all connections of the
client, so that [timestamp-seconds, sequence numbers] are never repeated
until the client restarts.

Sequence numbers can wraparound and get reset when the client restarts,
so they're used together with absolute timestamps server-side.  They
must be unique for at least 3 seconds, preferably 10 seconds, as they
guard against server time skew.

For 32-bit unsigned sequence numbers that are reset on client restart,
start at 1, and overflow from UINT32_MAX to 0, serialized as "seqnum" in
the CONNECT, MEASUREMENT and EVENTS messages, we declare the
"client-seqnum-v1" capability to the server.

Unsigned 32-bit sequence numbers are used because they're much easier to
get right [than unsigned 64-bit numbers] on 32-bit platforms (think
atomics or lack thereof), and fit both uint32_t an int64_t without
hassle, which helps with the 0.12.1 subset of the json-c API, as well as
other JSON libraries.