[MDC-1110] inetup: implement boot-time syncronism
Implement two ways to syncronize the client timestamps to system boot:
 * clock_getclock(CLOCK_BOOTTIME)
 * clock_getclock(CLOCK_MONOTONIC) + os_seconds_since_boot()

CLOCK_BOOTTIME exists in Linux and OpenBSD, only.  Even then, it is not
widely available in some embedded architectures where it was introduced
much later, and it might also be missing from some outdated embedded
libc.

Note that CLOCK_BOOTTIME is not always accelerated in the VDSO (depends
on how recent the kernel is).  It can be much slower to read than
CLOCK_MONOTONIC.

OTOH, CLOCK_MONOTONIC is usually VDSO-accelerated even on older kernels,
but it *stops* during platform suspend and thus loses sync with "seconds
since boot".  That requires periodic resync checks, each such check
requires a call to clock_gettime() and one to os_seconds_since_boot().
Should we detect that CLOCK_MONOTONIC is out of sync by more than two
seconds, we drop all connections, resync, and reconnect.

If we detect time went backwards on a resync -- which really should not
happen -- we enter unsync mode to warn the pipeline, and keep the old
sync offset, so as to not disturb the measurements too much.