[MDC-1090] inetup: suppress sequence numbers when uneeded
The network-availability-server only uses the client sequence number on
CONNECT messages, and doesn't need (or check for) such sequence numbers
on EVENTS or MEASUREMENT messages yet.

We can suppress the "seqnum" field on EVENTS and MEASUREMENT messages if
the server did not enable the "client-seqnum-v1" capability through a
ma_config message.

Note that we must not suppress sequence numbers in the CONNECT message
we send to the server, unless we are not announcing "client-seqnum-v1"
in that CONNECT message in the first place.