[MDC-882] inetup: suppress unwanted MEASUREMENT messages
Check whether the server reflected to us the msg-measurement capability,
and disable the sending of MEASUREMENT messages to that server if it did
not.  That's the proper way of using uptime2 capabilities, and it was an
error to not do so since day one of the msg-measurement feature.

Note that if the server sends a zero "measurement-period-seconds" in a
MA_CONFIG message to the client, it also disables measurements.  This
behavior must be kept for backwards compatibility with the older
versions of the client.