[MDC-1146] deb: disable expired DST Root CA X3 to avoid breakage
SIMET2 https-based APIs use Let's Encrypt certificates extensively.

On older Debian/Ubuntu-based systems with OpenSSL 1.0.x or old GNUTLS,
such as Ubuntu Xenial and Linux Educacional 6, one must *not* have the
expired "DST Root CA X3" active in the certificate store.  Otherwise,
Let's Encrypt trust chains could be rejected.

If it is still present in the certificate store, disable the "DST Root
CA X3 certificate" (expiration date in 2021-09-30) in the postinst
maintainer script, so that it won't cause issues.

While this is only required where openssl 1.0.x and older gnutls
versions are in use, removing expired CA certificates is not harmful, so
we can make it simple and just do it always.