[MDC-1149] simet-api: geo: persist cache across reboots (not on SIMETBox)
Define a SIMETVARDIR substitution parameter, for stuff that needs to go
in /var/lib/simet (non-ephemeral, variable data).

Define a SIMETCACHEDIR substitution parameter and set it either to
SIMETVARDIR (generic SIMET-MA), or SIMETRUNDIR (SIMETBox).

Use SIMETCACHEDIR to move the geolocation cache from /var/run to
/var/lib on non-SIMETBox builds.  This allows the cache to survive
reboots like the ones we're observing on thousands of MAs installed in
schools.  For SIMETBox, we leave it in var/run as we can't afford the
FLASH write cycles anyway.

The ABI with simet_geolocation_legado.sh is preserved since it uses
GEOLOC_CACHE.  However, the initscipts and other glue now need to ensure
that SIMETVARDIR exists and has the appropriate permissions.