[MDC-1149] simet-api: geo: add --max-age-spread parameter
Add --max-age-spread (and GEOLOC_MAXAGE_SPREAD config variable) to allow
for a stable spread (based on the system's agent-id) for geolocation
cache aging.

--max-age-spread should be specified in the same unit as --max-age (i.e.
seconds).  The same goes for GEOLOC_MAXAGE_SPREAD and GEOLOC_MAXAGE.

The spread will be within the interval -(max-age-spread / 2) to
+(max-age-spread / 2), centered on max-age.

This way we can spread the geolocation slot usage on the MA fleet over
time.  The MAs will try to geolocate when the cache is not yet present,
then renew it a first time after max-age modified by the spread, and
after that, it will renew at every max-age interval [unless their
agent-id changes].

One should use a *large* max-age to ensure a lower frequency of
geolocation usage, and a relatively smaller max-age-spread to avoid
overloading the target daily quota of geolocation requests.

Note that cache lifetime is configured in seconds, but for it to be of
any use as far as geolocation slot spreading is concerned, the MAs must
be more than 24 hours (86400 secons) apart.