[MDC-1151] deb: try much harder to stop lmapd on upgrades
Sometimes, we'd leave lmapd in a state where (dumb) lmapctl shutdown
wouldn't run, or something else would happen and it would survive a stop
signal.

When that happens, it becomes wedged at the next start under systemd, as
it (correctly) fails to start again -- it is already running -- and
systemd notices that.  However, futher attempts to stop it do not go
through because systemd thinks it is not started (which is *broken).
Whatever.  This happens at least on Debian 11.

Change the maintainer scripts to no longer care if the service stopped
or started properly.  This ensures we don't get any unattended-upgrade
updates wedged, so it ends up being much safer overall.  And it allows
us to recover from a wedged system.

Change the "stop" action for lmapd to SIGTERM everything named "lmapd"
under the nicbr-simet user, and to actually insist on it and SIGKILL any
that doesn't exit in 10s.  This will cover the vast majority of cases
where lmapd wouldn't stop, before.

Also direct dh to generate maintainer scripts that will always try to
stop the daemons on prerm, and start it again only in postinst.  It was
already doing that, but we might bump the "compat" level soon, and it
would be easy to forget that no, we *cannot* cope with the generally
unsafe, new *default* of "update a service's innads with it running".

As a final measure, run the new version of the "stop" action (hardcoded
in the postinst) under a dpkg --compare-version guard, when upgrading
from any versions of the package that did not have the strong "stop"
behavior.