[MDC-1151] deb: ensure init system knows we stopped lmapd
We need to make sure the init system knows we will stop lmapd if it is
still around.  And for the systemd compatibility layer, we also need to
ensure it has refreshed/regenerated the service unit.

So, call update-rc.d and invoke-rc.d stop on the service, *then*
forcefully kill any stray lmapds still around.

Otherwise, the service might not start at the end of postinst.