[MDC-1153] simet-api: retrieve, cache and log participant info
Add a participant info cache, AGENT_PARTICIPANT_INFO, that holds a JSON
object with the participants *attributes* when the MA is paired.

When unpaired, the cache will be empty.  It is an ephemeral file: if the
file is missing, the MA has not contacted the agent-registry yet since
boot.

When registering, retrieve participant information during the pairing
flow, and update the cache.  Output to the system log the name and
short_name attributes, should they be present and non-empty.