[MDC-1163] tcpbw: use incompressible payloads
By default, fill the buffer with data from /dev/urandom *once*, and use
that for bandwidth throughput measurement payloads.  This is intended to
make life a lot more difficult for "link optimizers" based on data
compression.

It is, however, far from complete a solution since we are *not* ensuring
the entire stream is uncompressible, nor that the contents of the
streams differ.

In fact, we will use the same random data sequence on every stream, for
now, in order to save on hotpath complexity and memory usage.

We also always begin sending from the start of the buffer, which means
the stream may end up far easier to compress than expected when the
kernel-side buffer is full.  That will be fixed in a future change.