[MDC-1163] tcpbw: do not always send from the start of the buffer
Keep track of how much of the sending buffer we have already used, so
that we send it as a (circular) stream instead of starting over from the
beggining at every send() syscall.

This didn't matter when we were sending zeroes, but it might matter when
we're sending random data to defeat link compression: consider the
situation where the kernel-side transmit buffer is nearly full: should
the kernel decide to wake us up (via select) asking for more data too
often, we'd end up creating an easily compressible pattern if we always
send from the start of the buffer.