[MDC-1165] simet_view_results: also ship for SIMETBox
It is useful to have simet_view_results.sh also on SIMETBox, so that we
can have the URLs being assembled in just one place: SIMETBox will be
modified to use the URL returned by simet_view_results as a base.

Add a new substitution (URLMATYPE) to handle the differing ma-type URL
parameter in simetbox.