[MDC-1170] simet-api: introduce an emergency schedule
For resiliency, it turns out we really need a fallback schedule in
place.  Otherwise, lmapd will refuse to start, and ends up in need of
heavy-handed healing to come back up (restore a schedule, reset the
crashloop counters of the init system supervisor, restart service).

Ship the emergency schedule "lmap-emergency-schedule.json" in the simet
lib directory.  It should be used through symlinks by the glue that
start/stop/restarts lmapd when it notices a schedule is missing.

The emergency schedule tries to get a proper schedule periodically.  It
could also try to register the MA, but that operation is currently not
carried out from within lmapd.

Note that the emergency schedule must contain the phrase
"ma-local_emergency-schedule" somewhere which grep can find (we use a
tag to hold it).  This allows simet*_lmap-fetch-schedule to replace it
with a locally generated schedule.