[MDC-1170] deb: use the emergency schedule
When the package is installed/updated, in simet-ma.postinst, instead of
just removing any previous schedule, replace it with the emergency
schedule.

On the init-scripts, symlink the emergency schedule into place when
there is no schedule in place yet.  This ensures we have something in
place.

This provides an schedule to lmapd should everything else fail,
otherwise it won't start.

Note: we used to do such schedule symlinking when we had a static
default schedule (but missed doing it in postinst at the time).  It was
removed by commit 9f70538f0e0bf0402999e828fcb9b3ae380b10f8, "[MDC-980]
deb: don't symlink the default schedule".  The new code is slightly
different.