[MDC-1198] simet-runner: rework agent-info context metric
Use the LMAP agent-info helper to generate the agent-info context
metric.  We keep our old code as a fallback should we fail to find the
agent-info LMAP helper.

Update the fallback code to actually return an error status if it fails,
instead of an empty table.  And fix the fallback code to not fail on
unimportant errors (only engine_name, engine_version and agent_family
are important).