[MDC-1250] twamp: send cookie on test packets
Send a cookie derived from session data negotiated with the server
through TWAMP_CONTROL, as part of the TWAMP_TEST payload.

The server can use that cookie to match TWAMP_TEST and TWAMP_CONTROL
connections when faced with insanely broken CGNAT (where UDP and TCP to
the same endpoint end up with differing source IPs).

Implement initially a very simplified cookie (version 1):
* 32-bit signature (must be sent in network byte order): 0x83b8c490
* 16-byte SID (as received from server in Accept-TW-Session)

Insert this at the start of the payload of the unauth-mode TWAMP_TEST
messages, for now.  Note that this will be relocated in the next change,
with an update to the 32-bit signature just in case, and that no servers
or clients were deployed to production with cookie signature 0x83b8c490.

If the server returns an all-zeroes SID, we disable the cookie
functionality.