[MDC-1250] twamp: RFC-6038 symmetric-mode padding
Insert a MBZ zone of 27 bytes before the "padding" (actually, as part of
padding right now), so that this area will always be zero, as per
RFC-6038 requirements for symmetric mode.

We don't currently implement RFC-6038 symmetric mode properly, but it is
better if our padding is always compatible with it.

This will move the SIMET cookie, and align it to a 4-byte boundary
inside a safer zone of the padding, keeping it away from the 27-byte MBZ
zone from RFC-6038.

Since this moved the cookie, update its signature to 0x83b8c493.  This
version of the cookie will be deployed to production on servers and
clients.

Note that this 27-byte MBZ zone can actually have some non-zero fields
inside it, e.g., RFC-7750, but only when both server and client agree,
otherwise interoperability issues with twamp-light and rfc-6038
symmetric mode will result.