simet_spoofer: try a lot more spoofed packets
Sometimes, simet-nping loses capture packets for some unexplained
reason.  If it happens at the wrong time, we could get a false negative
from the spoof test (i.e. think that no spoofed packets "made it" to
the destination).

We can't do parallel non-spoof and spoof at the same time yet (in which
case we would fail the test if the non-spoofed packets fail to arrive),
but we can try 25 probes instead of 5 to give ourselves more chances of
escaping periods where the packet capture in simet-nping is failing.

This is a temporary workaround.