inetup: tcpaq: split from main source
Split the tcpaq (tcp message queue) layer from the main executable, as
it will become common code sooner or later.

Introduce a new struct, struct tcpaq_conn, which holds the socket fd as
well as the input and output message buffers.

struct tcpaq_conn is typically embedded in some other struct, so it is
not dynamically allocated or freed by the tcpaq layer.

There is a tcpaq_init() to initalize it and allocate the buffers, and a
tcpaq_free_members() to dispose of anything inside struct tcpaq_conn
(but not the struct tcpaq_conn itself).

Other layers (such as the message layer itself) are expected to poke
inside struct tcpaq_conn a bit, at least for now.