[MDC-1310] twampc: really implement connection timeout cli parameter
When refactoring the code to allow for a per-packet timeout, the -t
parameter was repurposed to be a "connection timeout".  Unfortunately,
it was not really enabled in the code during the refactoring, and a
hardcoded value of 2 seconds remained in place.

Change the code to use the value of the -t parameter as the TCP
connection timeout.  UDP "connections" have no timeout (they either
fail, or succeed immediately), but we use the same timeout for simmetry.

The default TCP connection timeout is now 15s as it was supposed to be,
and the -t command line parameter can change it as it was supposed to.

Fixes: [MDC-419] simet-ma: twamp: per-packet timeout, timeouts in us