[MDC-201] twamp: refactor timestamp handling
Clean up timestamp handling to use less pointers, we will need this to
address fields in packed structures.  To do that, move most trivial
conversions to the header, as static inline functions.

It should not result in much worse code as long as the optimizing
compiler does a moderately competent job of it.

Add a timestamp_to_microsec(), as report.c actually wants that.