[MDC-201] twamp: configurable payload size
Add the -s option to configure the payload size.  We accept payload
sizes 128 to 65400 in the client.

Note that the server will refuse sessions outside of its service range,
which is likely to be far more restrictive than this interval.

The default size is unchanged (128 bytes of payload, 170-byte packet).