[MDC-1321] twamp: implement RFC6038 symmetric mode
Request symmetric mode from the server (if it offers that option), since
we discard any packets that have a size that doesn't match what we sent.

Old SIMET2 TWAMP reflectors/servers always operate in symmetric mode,
but do not implement the option properly, so we just continue with an
warning even if the server does not offer symmetric mode.  It is
relevant that the perfSonar TWAMP reflector does the same as our older
TWAMP reflector, so it looks like we will not do more than warn anytime
soon.

Our payload should be compatible with RFC6038 already: it is either
all-zeroes, or it has the 27-octet MBZ before non-zero content.

Warning: we set the padding size as if that 27-octet MBZ was part of the
padding.  The RFCs are unclear on this point, and this could be an
interoperability issue.  For what little is it worth, we're compatible
with perfSonar, but perfSonar *also* does symmetric mode *without*
full RFC6038 support server/reflector-side, and their client does not
ask for symmetric mode when offered.