[MDC-1322] twamp: pass by reference parameters to twamp_test()
Use a reference (const pointer to non-const data) to pass the
TestParameters struct to twamp_test(), since it contains pointers to
other structures that *are* modified by twamp_test(), and we use the
modified contents in the caller...

This is at least a lot less confusing and more idiomatic to C
programmers, which should avoid a bug or two in the long run.

It is still asking for a major refactor to use a full context
everywhere, and have all data ownership documented, though.