[MDC-1322] twamp: implement source address binding
For TWAMP Light, it is quite often required to bind the source address
of TWAMP-TEST to a specific IP and/or UDP port.

In order to implement that functionality:
* Refactor ip_family support to use sa_family_t AF_UNSPEC,INET,INET6 instead
  of integers 0, 4 and 6.
* Refactor and/or rename some helper functions dealing with struct sockaddr
* Add a ss_source field to TWAMPParameters, and bind() to it when non-NULL
* Refactor TWAMPParameters a bit to const'ify things
* Fix a few minor memory leaks

To set the source address or port (or both), there is a new "-I
[<source_ip>][:<port>]" command line option:

  -I 127.0.0.1
  -I :8642
  -I 127.0.0.1:8642
  -I [::1]:9999

For TWAMP-Light, -I affects TWAMP-TEST.  For TWAMP, it currently does
not affect anything, but futher refactors on twamp_run_client() might
change that.