[MDC-1325] twamp: do not leak thread on error in twamp_test()
The sender thread was not always joining the receiver thread on error,
and would leak memory allocated by pthread_create() or the receiver
thread itself.

In practice, it would only happen if send() failed as the other error
conditions are extremely rare.

Since the client exits at the moment, a cleanup is done soon enough, but
this would be a far more serious issue in the future should we start
doing twamp_test() in a loop, or convert it into a library in a
long-lived application.