[MDC-1323] twamp: stop threads on error
When one of the threads hits an error condition, the program may end up
waiting for the other thread to finish needlessly (and that might even
mean *sending packets* that won't be received...).

And it can take a lot of time to stop, depending on the test parameters,
such as packet count, or intervals.

Add a volatile flag that, when set, causes the threads to stop, and set
it when either thread hits an error condition.