[MDC-199] add timespec_lt(), and a comment to timestamp.c/h
We will need to compare two timespecs, so add a inline static helper to
do it in timestamp.h, and while at it add a comment related to timespec
normalization in timestamp.c.

Note that POSIX allows struct timespec tv_nsec in [-1, 10E9) but only in
very special circunstances. tv_nsec better be non-negative if you do not
want to risk incorrect behavior.