[MDC-199] twamp: change systematic send error into jitter
Switch to (where available) clock_nanosleep() with TIMER_ABSTIME, which
changes some of the the systematic error we had on the packet send
schedule (always a bit more than the desired interval, thus further
deviating from the desired packet train) into sending jitter.

With this change, it is possible for the interpacket delay in the packet
train to be a bit less than the desired target, instead of always being
a bit more than the desired target.