[MDC-1322] twamp: stop early when only corrupt packets are received
If the reflector is misbehaving to the point it is likely NOT a TWAMP
reflector, we should stop sending packets.  In twamp-light mode, one
could easily pester some UDP-based services that reflect packets, such
as a DNS resolver.

For all TWAMP modes:

* Stop a test run early if we are only receiving corrupted packets.

* Dectect as corrupt any reflected packets that have sequence numbers
  outside of the valid range.

We also check for a all-zeroes sender timestamp, and flag that as
corrupt since we should never send those.  We could be more strict than
that, but at that point we'd be better off switching to autenticated
mode.

Note: we can't check reflector timestamps.  We can't assume anything
about the reflector's clocks -- that's asking for trouble -- and at
least one big-name chinese network device vendor sets "Received
Timestamp" and "Timestamp" fields to all-zeroes in the reflected
packet...