[MDC-1251] twamp: restore MSG_TRUNC for recv()
recv(... MSG_TRUNC || MSG_DONTWAIT) results in recv(... 1), instead of
setting both MSG_TRUNC and MSG_DONTWAIT bits.  For Linux x86, 1 is the
same as MSG_OOB.

Fix it to be recv(... MSG_TRUNC | MSG_DONTWAIT) as it should be.

Fortunatey, the loss of MSG_TRUNC and MSG_DONTWAIT doesn't make much of
a difference in practice, as the socket is already in non-blocking mode
and we only use MSG_TRUNC to reject unexpectedly long packets.

We shall keep the MSG_DONTWAIT in there as it is actually safer
maintenance-wise (now that the bug is fixed) against further breakage,
even if it was its superfluous addition that allowed the bug to creep
in...

Fixes: commit 8f26f373841f1287de4b79190647ae805ec8b81d
       [MDC-1251] twamp: minor refactoring for TWAMP_TEST receive