[MDC-1329] simet-api: simetbox: do not trust openwrt initscripts
We cannot trust OpenWRT to keep their act together re. non-root, so
bypass their hackjob entirely.

We already validate the lmapd config before we attempt a simet-lmapd
reconfig (which is a restart and WILL result in a non-running lmapd if
the config is invalid/broken), so just call "lmapctl reload" directly.