[MDC-1335] tcpbw: refactor and fix receiveDownloadPackets()
Refactor receiveDownloadPackets() to use pselect and CLOCK_MONOTINIC,
and also change the way it schedules the sampling of results.

Before this refactoring, receiveDownloadPackets() would wait for at
least <sampling period> after taking a sample.  This generates a
systematic error that pretty much ensured the code would never manage to
take the last sample before the measurement window closed, among other
problems.

Also, receiveDownloadPackets() was timestamping the clock *before*
calling pselect and the recv() loop.  However, pselect() can wait for
quite a while (the entire sampling period, in fact) if there is no data
available to be read, so the error in the sampling period we would
record in the report would have a large error margin.

So, for a 11 second measurement, sampled every 500ms, we would only get
at most 21 samples (instead of 22).

Now, we advance the sampling window in fixed steps if <sample period>, so
we might wait less than <sampling period> between samples if the last
sampling was done "late".  The clock is timestamped only after pselect()
returns, so pselect() is no longer contributing to measurement error
margins, and the outer loop was restructured so that the code has a much
better chance of taking that last sample before noticing the measurement
window closed.

Actually, we timestamp after both pselect() and recv() are done, as
that seems to work a little better.  Doing it after recv() is not nearly
as important as doing it after the pselect(), though.

Also, switch from round-to-zero to round-to-nearest when converting the
timestamp interval from nanoseconds to miliseconds.  Otherwise we have a
larger margin of error in the interval we add to the report.