[MDC-1335] tcpbw: add timespec.h to prepare for refactoration
We shall change a lot of struct timeval to struct timespec in the
following commits, and we will need some common code to deal with those.

Add it to a new header, timespec.h.

Note that the timespec functions are very short, and declared "inline",
so it will be optimized by the compiler into decent code folded into the
caller, while still being more type-safe than non-extremely-convoluted
preprocessor macros.