[MDC-199] twamp: round to nearest when converting to/from ntpv3 timestamps
We use floating point math to convert to/from ntp v3 timestamps.  Take
care to round to the nearest integer, instead of doing an implicit
floor() by simple typecasting.

Since we're dealing with positive numbers only, it is enough to sum 0.5
to the floating point intermediary result before truncating it to
integer (i.e. before doing the implicit floor()).