inetup: fix misuse of field length with fscanf %ms
This addresses a harmless warning from CLANG.  There is no risk of a
buffer overflow because the libc will not suddenly write a pointer that
is larger than (sizeof(char*)) just becase the format specifier tells it
that it could.

However, it is probably a good idea to limit that agent-id size to
something sane, and ignore it as invalid if it is too large, so add such
limits properly.

Refuse files with more than 256KiB, and refuse an agent-id that is too
large for what we're using them right now.